requirements.txt
# AI Newsletter SaaS Requirements
requests==2.31.0
feedparser==6.0.10
transformers==4.35.2
torch==2.1.1
flask==3.0.0
python-dotenv==1.0.0
numpy==1.24.3
scikit-learn==1.3.2
beautifulsoup4==4.12.2
lxml==4.9.3
jinja2==3.1.2
markupsafe==2.1.3
gunicorn==21.2.0